Readme for AlmaLinux 8.6

+++ alma/common/hpc-tuning.sh
@@ -15,7 +15,45 @@ EOF
 
 # Enable reclaim mode
 echo "vm.zone_reclaim_mode = 1" >> /etc/sysctl.conf
+echo "net.ipv4.neigh.default.gc_thresh1 = 4096" >> /etc/sysctl.conf
+echo "net.ipv4.neigh.default.gc_thresh2 = 8192" >> /etc/sysctl.conf
+echo "net.ipv4.neigh.default.gc_thresh3 = 16384" >> /etc/sysctl.conf
+echo "sunrpc.tcp_max_slot_table_entries = 128" >> /etc/sysctl.conf
+
+## Systemd service for starting sunrpc and adding setting parameters
+cat <<EOF >/usr/sbin/sunrpc_tcp_settings.sh
+#!/bin/bash
+
+modprobe sunrpc
 sysctl -p
+EOF
+
+chmod 755 /usr/sbin/sunrpc_tcp_settings.sh
+
+cat <<EOF >/etc/systemd/system/sunrpc_tcp_settings.service
+[Unit]
+Description=Set sunrpc tcp settings
+
+[Service]
+User=root
+Type=oneshot
+ExecStart=/usr/sbin/sunrpc_tcp_settings.sh
+RemainAfterExit=true
+
+[Install]
+WantedBy=multi-user.target
+EOF
+
+systemctl enable sunrpc_tcp_settings
+systemctl start sunrpc_tcp_settings
+systemctl is-active --quiet sunrpc_tcp_settings
+
+error_code=$?
+if [ ${error_code} -ne 0 ]
+then
+    echo "sunrpc_tcp_settings service inactive!"
+    exit ${error_code}
+fi
 
 # Remove auoms if exists - Prevent CPU utilization by auoms
 if yum list installed azsec-monitor >/dev/null 2>&1; then yum remove -y azsec-monitor; fi
@@ -24,13 +62,15 @@ if yum list installed azsec-monitor >/dev/null 2>&1; then yum remove -y azsec-mo
 yum update -y WALinuxAgent
 
 # Configure WALinuxAgent
-sudo sed -i -e 's/# OS.EnableRDMA=y/OS.EnableRDMA=y/g' /etc/waagent.conf
-echo "Extensions.GoalStatePeriod=300" | sudo tee -a /etc/waagent.conf
-echo "Extensions.InitialGoalStatePeriod=6" | sudo tee -a /etc/waagent.conf
-echo "OS.EnableFirewallPeriod=300" | sudo tee -a /etc/waagent.conf
-echo "OS.RemovePersistentNetRulesPeriod=300" | sudo tee -a /etc/waagent.conf
-echo "OS.RootDeviceScsiTimeoutPeriod=300" | sudo tee -a /etc/waagent.conf
-echo "OS.MonitorDhcpClientRestartPeriod=60" | sudo tee -a /etc/waagent.conf
-echo "Provisioning.MonitorHostNamePeriod=60" | sudo tee -a /etc/waagent.conf
-sudo systemctl restart waagent
+sed -i -e 's/# OS.EnableRDMA=y/OS.EnableRDMA=y/g' /etc/waagent.conf
+cat << EOF | tee -a /etc/waagent.conf
+Extensions.GoalStatePeriod=300
+Extensions.InitialGoalStatePeriod=6
+OS.EnableFirewallPeriod=300
+OS.RemovePersistentNetRulesPeriod=300
+OS.RootDeviceScsiTimeoutPeriod=300
+OS.MonitorDhcpClientRestartPeriod=60
+Provisioning.MonitorHostNamePeriod=60
+EOF
+systemctl restart waagent
 $COMMON_DIR/write_component_version.sh "WAAGENT" $(waagent --version | head -n 1 | awk -F' ' '{print $1}' | awk -F- '{print $2}')
